requirements.txt
aws-cdk-lib==2.108.1
aws-cdk.aws-glue-alpha>=2.108.1-alpha.0
constructs>=10.0.0,<11.0.0
